pyproject.toml
[tool.poetry]
name = "agentartificial"
version = "0.1.0"
description = ""
authors = ["Richard <bakobiibizo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.14.2"
python-dotenv = "^1.0.1"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
requests = "^2.31.0"
loguru = "^0.7.2"
pytest = "^8.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
